convert to multi module maven project
