+++ jenkins-plugin/src/main/resources/io/jenkins/plugins/remotingopentelemetry/RemotingOpenTelemetryConfiguration/config.jelly
