+++ monitoring-engine/src/main/java/io/jenkins/plugins/remotingopentelemetry/engine/EngineConfiguration.java
